pyproject.toml
[tool.poetry]
name = "extrack"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pysimplegui = "^5.0.4"
black = "^24.4.2"
ddt = "^1.7.2"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
